requirements.txt
Django>=3.1.4,<3.2.0
djangorestframework>=3.12.2,<3.13.0
psycopg2>=2.8.6,<2.9.0
django-cors-headers>=3.6.0,<3.8.0
Pillow>=8.1.0,<8.2.0
uWSGI>=2.0.18,<2.1
